added a dictionar in js to handle translations on elements that change on click, fx. skrá i svef btn
