--- src/assets/js/partials/component-translate.js
@@ -1,14 +1,0 @@
-const TranslateLanguages = {
-  languageSwitch(string, lang) {
-    return this.dictionary[string][lang];
-  },
-  dictionary: {
-    required: { is: 'skilyrt', en: 'required' },
-    skilyrt: { is: 'skilyrt', en: 'required' }
-  },
-  currentLanguage() {
-    return myAjax.currentLang
-  }
-}
-
-module.exports = TranslateLanguages
